requirements.txt
RPi.GPIO
pyzmq